+++ firmware/hw_layer/digital_input/trigger/trigger_input_exti.cpp
@@ -42,7 +42,10 @@ static void cam_callback(void *arg, efitick_t stamp) {
 
 	bool rise = (palReadLine(pal_line) == PAL_HIGH);
 
-	if (rise ^ engineConfiguration->invertCamVVTSignal) {
+	int camIndex = CAM_BY_INDEX(index);
+	bool invertSetting = camIndex == 0 ? engineConfiguration->invertCamVVTSignal : engineConfiguration->invertExhaustCamVVTSignal;
+
+	if (rise ^ invertSetting) {
 		hwHandleVvtCamSignal(TriggerValue::RISE, stamp, index);
 	} else {
 		hwHandleVvtCamSignal(TriggerValue::FALL, stamp, index);